requirements.txt
# Django and Core Dependencies
django==5.0.7
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
requests==2.31.0

# Django REST Framework for APIs
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Asynchronous Task Processing
celery==5.4.0
redis==5.0.4

# Frontend and Templating
django-htmx==1.17.2

# Reporting and File Generation
openpyxl==3.1.2
WeasyPrint==62.3

# Development and Debugging
django-debug-toolbar==4.4.1

# Utilities
Pillow==10.3.0
Faker==25.2.0
tqdm==4.66.4 # <--- تمت إضافة هذه المكتبة لشريط التقدم